Remove Titan wording from LICENSE.txt [skip ci]
This file was copied from the main JanusGraph repo, but since the FoundationDB
was not derived from Titan, we don't need the wording and timeline around the
Titan fork, as all code in this repo is under the Apache 2.0 license, while all
documentation is under the CC-BY-4.0 license, since it all post-dates the Titan
fork in 2016.

Also add APACHE-2.0.txt and CC-BY-4.0.txt with verbatim copies of those two
licenses as specified in LICENSE.txt.

Closes #39

Signed-off-by: Misha Brukman <mbrukman@google.com>